GH-3483: make the Federation Strategy Factory configurable
This change allows to supply a custom FederationEvaluationStrategy
through the FedXFactory.

The previous mechanism of defining the class names trough the FedXConfig
and using reflection for instantiation is removed. A corresponding note
is added to the changelog
