+++ site/content/release-notes/4.0.0.md
@@ -204,6 +204,13 @@ that use their own implementations of the `IRI` interface will need to verify
 and possibly adjust their `hashCode()` and `equals()` implementation to conform
 to this.
 
+### Refined extension points for evaluation strategies in federation
+
+Removed support for defining `sparqlEvaluationStrategy` and `sailEvaluationStrategy` using `FedXConfig`.
+
+Use `FedXFactory#withFederationEvaluationStrategyFactory` instead to supply a `FederationEvaluationStrategyFactory`.
+
+
 ## Acknowledgements
 
 This release was made possible by contributions from Andreas Schwarte, Florian
