+++ tools/federation/src/main/java/org/eclipse/rdf4j/federated/FedXConfig.java
@@ -11,11 +11,9 @@ import java.util.Optional;
 
 import org.eclipse.rdf4j.federated.cache.SourceSelectionCache;
 import org.eclipse.rdf4j.federated.cache.SourceSelectionMemoryCache;
-import org.eclipse.rdf4j.federated.evaluation.FederationEvalStrategy;
-import org.eclipse.rdf4j.federated.evaluation.SailFederationEvalStrategy;
-import org.eclipse.rdf4j.federated.evaluation.SparqlFederationEvalStrategy;
 import org.eclipse.rdf4j.federated.evaluation.concurrent.ControlledWorkerScheduler;
 import org.eclipse.rdf4j.federated.evaluation.concurrent.TaskWrapper;
+import org.eclipse.rdf4j.federated.evaluation.iterator.ConsumingIteration;
 import org.eclipse.rdf4j.federated.monitoring.QueryLog;
 import org.eclipse.rdf4j.federated.monitoring.QueryPlanLog;
 import org.eclipse.rdf4j.federated.write.DefaultWriteStrategyFactory;
@@ -58,10 +56,6 @@ public class FedXConfig {
 
 	private String sourceSelectionCacheSpec = null;
 
-	private Class<? extends FederationEvalStrategy> sailEvaluationStrategy = SailFederationEvalStrategy.class;
-
-	private Class<? extends FederationEvalStrategy> sparqlEvaluationStrategy = SparqlFederationEvalStrategy.class;
-
 	private Class<? extends WriteStrategyFactory> writeStrategyFactory = DefaultWriteStrategyFactory.class;
 
 	private TaskWrapper taskWrapper = null;
@@ -99,36 +93,6 @@ public class FedXConfig {
 	 */
 	public FedXConfig withLogQueries(boolean flag) {
 		this.isLogQueries = flag;
-		return this;
-	}
-
-	/**
-	 * Set the {@link FederationEvalStrategy} for SPARQL federations. See {@link #getSPARQLEvaluationStrategy()}.
-	 *
-	 * <p>
-	 * Can only be set before federation initialization.
-	 * </p>
-	 *
-	 * @param sparqlEvaluationStrategy
-	 * @return the current config
-	 */
-	public FedXConfig withSparqlEvaluationStrategy(Class<? extends FederationEvalStrategy> sparqlEvaluationStrategy) {
-		this.sparqlEvaluationStrategy = sparqlEvaluationStrategy;
-		return this;
-	}
-
-	/**
-	 * Set the {@link FederationEvalStrategy} for SAIL federations. See {@link #getSailEvaluationStrategy()}.
-	 *
-	 * <p>
-	 * Can only be set before federation initialization.
-	 * </p>
-	 *
-	 * @param sailEvaluationStrategy
-	 * @return the current config
-	 */
-	public FedXConfig withSailEvaluationStrategy(Class<? extends FederationEvalStrategy> sailEvaluationStrategy) {
-		this.sailEvaluationStrategy = sailEvaluationStrategy;
 		return this;
 	}
 
@@ -450,32 +414,6 @@ public class FedXConfig {
 	 */
 	public String getSourceSelectionCacheSpec() {
 		return this.sourceSelectionCacheSpec;
-	}
-
-	/**
-	 * Returns the class of the {@link FederationEvalStrategy} implementation that is used in the case of SAIL
-	 * implementations, e.g. for native stores.
-	 * <p>
-	 * Default {@link SailFederationEvalStrategy}
-	 * </p>
-	 *
-	 * @return the evaluation strategy class
-	 */
-	public Class<? extends FederationEvalStrategy> getSailEvaluationStrategy() {
-		return sailEvaluationStrategy;
-	}
-
-	/**
-	 * Returns the class of the {@link FederationEvalStrategy} implementation that is used in the case of SPARQL
-	 * implementations, e.g. SPARQL repository or remote repository.
-	 * <p>
-	 * Default {@link SparqlFederationEvalStrategy}
-	 * </p>
-	 *
-	 * @return the evaluation strategy class
-	 */
-	public Class<? extends FederationEvalStrategy> getSPARQLEvaluationStrategy() {
-		return sparqlEvaluationStrategy;
 	}
 
 	/**
