+++ tools/federation/src/main/java/org/eclipse/rdf4j/federated/evaluation/FederationEvaluationStrategyFactory.java
@@ -7,7 +7,7 @@
  *******************************************************************************/
 package org.eclipse.rdf4j.federated.evaluation;
 
-import org.eclipse.rdf4j.federated.FedXConfig;
+import org.eclipse.rdf4j.federated.FedXFactory;
 import org.eclipse.rdf4j.federated.FederationContext;
 import org.eclipse.rdf4j.federated.FederationManager.FederationType;
 import org.eclipse.rdf4j.query.Dataset;
@@ -16,8 +16,22 @@ import org.eclipse.rdf4j.query.algebra.evaluation.impl.EvaluationStatistics;
 import org.eclipse.rdf4j.query.algebra.evaluation.impl.StrictEvaluationStrategyFactory;
 
 /**
- * Factory class for retrieving the {@link FederationEvalStrategy} to be used
+ * Factory class for retrieving the {@link FederationEvalStrategy} to be used.
+ * 
+ * <p>
+ * Default strategies:
+ * </p>
+ * 
+ * <ul>
+ * <li>local federation: {@link SailFederationEvalStrategy}</li>
+ * <li>endpoint federation: {@link SparqlFederationEvalStrategy}</li>
+ * <li>hybrid federation: {@link SparqlFederationEvalStrategy}</li>
+ * </ul>
  *
+ * <p>
+ * Customized strategies can be supplied to the federation using
+ * {@link FedXFactory#withFederationEvaluationStrategyFactory(FederationEvaluationStrategyFactory)}
+ * 
  * @author Andreas Schwarte
  */
 public class FederationEvaluationStrategyFactory extends StrictEvaluationStrategyFactory {
@@ -51,41 +65,14 @@ public class FederationEvaluationStrategyFactory extends StrictEvaluationStrateg
 			EvaluationStatistics evaluationStatistics) {
 
 		// Note: currently dataset, triplesource and statistics are explicitly ignored in the federation
-		return getEvaluationStrategy(federationType, federationContext);
-	}
-
-	/**
-	 * Return an instance of {@link FederationEvalStrategy} which is used for evaluating the query. The type depends on
-	 * the {@link FederationType} as well as on the actual implementations given by the configuration, in particular
-	 * this is {@link FedXConfig#getSailEvaluationStrategy()} and {@link FedXConfig#getSPARQLEvaluationStrategy()}.
-	 *
-	 * @param federationType
-	 * @param federationContext
-	 * @return the {@link FederationEvalStrategy}
-	 */
-	private static FederationEvalStrategy getEvaluationStrategy(FederationType federationType,
-			FederationContext federationContext) {
 
 		switch (federationType) {
 		case LOCAL:
-			return instantiate(federationContext.getConfig().getSailEvaluationStrategy(), federationContext);
+			return new SailFederationEvalStrategy(federationContext);
 		case REMOTE:
 		case HYBRID:
 		default:
-			return instantiate(federationContext.getConfig().getSPARQLEvaluationStrategy(), federationContext);
-		}
-	}
-
-	private static FederationEvalStrategy instantiate(Class<? extends FederationEvalStrategy> evalStrategyClass,
-			FederationContext federationContext) {
-		try {
-			return (FederationEvalStrategy) evalStrategyClass
-					.getDeclaredConstructor(FederationContext.class)
-					.newInstance(federationContext);
-		} catch (InstantiationException e) {
-			throw new IllegalStateException("Class " + evalStrategyClass + " could not be instantiated.", e);
-		} catch (Exception e) {
-			throw new IllegalStateException("Unexpected error while instantiating " + evalStrategyClass + ":", e);
+			return new SparqlFederationEvalStrategy(federationContext);
 		}
 	}
 }
